Implements forEach
This is needed to make can-stache for/of work.
